refactor: extract navigation.js (breadcrumb + copy-path) from index.js
- navigation.js: setupBreadcrumbs() — file-tree fetching, dropdown
  construction, expand/collapse behaviour, copy-path button
- index.js: 785 → 625 lines (160 removed)
- Remove now-unused propagateFetchError/nodeIs/postForm imports from
  index.js (they moved to navigation.js)
